Added custom time setup to give hour, minute and seconds to
send messages after given time.

- Improved update downloader
- Removed warning counter
- Now the json not sends the json format to chat

+++ src/hu/montlikadani/AutoMessager/bukkit/UpdateDownloader.java
@@ -25,10 +25,11 @@ public class UpdateDownloader {
 		}
 
 		String msg = "";
+		String versionString = "";
 		String lineWithVersion = "";
 
-		double newVersion = 0d;
-		double currentVersion = 0d;
+		int newVersion = 0;
+		int currentVersion = 0;
 
 		try {
 			URL githubUrl = new URL("https://raw.githubusercontent.com/montlikadani/AutoMessager/master/plugin.yml");
@@ -42,28 +43,27 @@ public class UpdateDownloader {
 				}
 			}
 
-			String[] nVersion;
-			String[] cVersion;
+			versionString = lineWithVersion.split(": ")[1];
+			String nVersion = versionString.replaceAll("[^0-9]", "");
+			newVersion = Integer.parseInt(nVersion);
 
-			String versionString = lineWithVersion.split(": ")[1];
-			nVersion = versionString.replaceAll("[^0-9.]", "").split("\\.");
-			newVersion = Double.parseDouble(nVersion[0] + "." + nVersion[1]);
-
-			cVersion = AutoMessager.getInstance().getDescription().getVersion().replaceAll("[^0-9.]", "").split("\\.");
-			currentVersion = Double.parseDouble(cVersion[0] + "." + cVersion[1]);
+			String cVersion = AutoMessager.getInstance().getDescription().getVersion().replaceAll("[^0-9]", "");
+			currentVersion = Integer.parseInt(cVersion);
 
 			if (newVersion > currentVersion) {
-				if (sender.equals("player")) {
+				if ("player".equals(sender)) {
 					msg = "&8&m&l--------------------------------------------------\n"
 							+ "&aA new update for AutoMessager is available!&4 Version:&7 " + versionString
 							+ (conf.getBoolean("download-updates") ? ""
 									: "\n&6Download:&c &nhttps://www.spigotmc.org/resources/43875/")
 							+ "\n&8&m&l--------------------------------------------------";
-				} else if (sender.equals("console")) {
+				} else if ("console".equals(sender)) {
 					msg = "New version (" + versionString
 							+ ") is available at https://www.spigotmc.org/resources/43875/";
 				}
-			} else if (sender.equals("console")) {
+
+				msg = Util.colorMsg(msg);
+			} else if ("console".equals(sender)) {
 				return "You're running the latest version.";
 			}
 
@@ -94,6 +94,12 @@ public class UpdateDownloader {
 						}
 
 						File jar = new File(updatesFolder + per + name + ".jar");
+						if (jar.exists()) {
+							in.close();
+							cancel();
+							return;
+						}
+
 						Files.copy(in, jar.toPath(), StandardCopyOption.REPLACE_EXISTING);
 
 						in.close();
